# Speed up concurrency state implementation