<!-- Please include the 'why' behind your changes if no issue exists -->
## Proposed Changes

**Note:** Please take your time reviewing this. I might totally have missed a race condition somewhere and this might be utter garbage. If so, I'm super happy to bin this. This was mostly some friday "let's solve a trickyish concurrency problem quicker" type fun :joy:.

This rewrites the ConcurrencyStateHandler from a channel based approach
into an atomic/mux based approach, to minimize contention and allow for
more concurrent processing.

An RWMutex is used to make sure that calling either pause or resume is
mutually exclusive to each other and also mutually exclusive to the
requests themselves, though the requests can happily run in parallel.

An important race to consider is a request entering into the pausing
logic while another request enters the resume block. Before pausing,
it's important to recheck if `inFlight` is still 0 **after** obtaining
the lock to make sure no other request has raced obtaining the lock. The
race can happen because the increment/decrement of `inFlight` is not
atomic to getting the lock so the following order of operations could
happen:

1. A: inFlight.Dec() == 0 --> true
2. B: inFlight.Inc()
3. B: mux.RLock()
4. B: !paused --> true
5. A: !paused --> true
6. A: pause()
7. B: h.ServeHTTP --> Boom.

In step 6, the pause would not have been excecuted in the actual
implementation though, because we doublecheck the state of `inFlight`,
which wouldn't have been 0 anymore.

## Benchmark diff

```
benchmark                                                                             old ns/op     new ns/op     delta
BenchmarkConcurrencyStateProxyHandler/sequential-breaker-10-no-reports-16             8123          1770          -78.21%
BenchmarkConcurrencyStateProxyHandler/parallel-breaker-10-no-reports-16               1919          672           -64.98%
BenchmarkConcurrencyStateProxyHandler/sequential-breaker-infinite-no-reports-16       8093          1319          -83.70%
BenchmarkConcurrencyStateProxyHandler/parallel-breaker-infinite-no-reports-16         1886          631           -66.55%
BenchmarkConcurrencyStateProxyHandler/sequential-breaker-10-many-reports-16           7657          1854          -75.79%
BenchmarkConcurrencyStateProxyHandler/parallel-breaker-10-many-reports-16             1889          693           -63.34%
BenchmarkConcurrencyStateProxyHandler/sequential-breaker-infinite-many-reports-16     7149          1428          -80.03%
BenchmarkConcurrencyStateProxyHandler/parallel-breaker-infinite-many-reports-16       1839          673           -63.38%

benchmark                                                                             old allocs     new allocs     delta
BenchmarkConcurrencyStateProxyHandler/sequential-breaker-10-no-reports-16             11             10             -9.09%
BenchmarkConcurrencyStateProxyHandler/parallel-breaker-10-no-reports-16               11             10             -9.09%
BenchmarkConcurrencyStateProxyHandler/sequential-breaker-infinite-no-reports-16       7              6              -14.29%
BenchmarkConcurrencyStateProxyHandler/parallel-breaker-infinite-no-reports-16         7              6              -14.29%
BenchmarkConcurrencyStateProxyHandler/sequential-breaker-10-many-reports-16           11             10             -9.09%
BenchmarkConcurrencyStateProxyHandler/parallel-breaker-10-many-reports-16             11             10             -9.09%
BenchmarkConcurrencyStateProxyHandler/sequential-breaker-infinite-many-reports-16     7              6              -14.29%
BenchmarkConcurrencyStateProxyHandler/parallel-breaker-infinite-many-reports-16       7              6              -14.29%

benchmark                                                                             old bytes     new bytes     delta
BenchmarkConcurrencyStateProxyHandler/sequential-breaker-10-no-reports-16             912           816           -10.53%
BenchmarkConcurrencyStateProxyHandler/parallel-breaker-10-no-reports-16               912           816           -10.53%
BenchmarkConcurrencyStateProxyHandler/sequential-breaker-infinite-no-reports-16       704           608           -13.64%
BenchmarkConcurrencyStateProxyHandler/parallel-breaker-infinite-no-reports-16         704           608           -13.64%
BenchmarkConcurrencyStateProxyHandler/sequential-breaker-10-many-reports-16           912           816           -10.53%
BenchmarkConcurrencyStateProxyHandler/parallel-breaker-10-many-reports-16             912           816           -10.53%
BenchmarkConcurrencyStateProxyHandler/sequential-breaker-infinite-many-reports-16     704           608           -13.64%
BenchmarkConcurrencyStateProxyHandler/parallel-breaker-infinite-many-reports-16       704           608           -13.64%
```

/assign @psschwei @julz