@markusthoemmes 2021-09-03
/hold

Actually found a race that's not handled properly