@markusthoemmes 2021-09-03
Almost sure I can actually replace the waiters stuff with a doubly checked lock and allowing the requests to race to resume 🤔