@psschwei 2021-09-03
Interesting... I'm going to check how this plays with the pause/resume functions from #11802 (which is awaiting review ;) ). I don't anticipate any issues, just want to poke around with it a bit. Then will give a review, probably after the US holiday weekend. 

(The performance improvement is quite nice!)