@psschwei 2021-09-04
One interesting thing I noticed is that, in certain cases, the pause/resume functions are making two HTTP calls. 

Here's what I did:

* I cloned your branch and added pause/resume functions from #11802 (and made some minor changes to function signatures): https://github.com/psschwei/serving/tree/rewrite-concurrencystate
* Set up a dev environment on minikube using this branch (w/ kourier networking)
* Deployed the [pod-freezer](https://github.com/psschwei/pod-freezer) service, which as of now just logs when a pause/resume request comes in
* Deployed julz's [sleeptalker app](https://github.com/psschwei/pod-freezer), which takes two seconds to run its main task (and thus doesn't result in pause being called immediately after resume)
* Used hey to simulate multiple requests: `hey -c 7 -n 10 http://example.default.10.98.123.235.sslip.io`

Here's the interesting part:

logs for queue-proxy:
```
{"severity":"INFO","timestamp":"2021-09-04T00:19:07.441246031Z","logger":"queueproxy","caller":"queue/concurrency_state.go:79","message":"Requests increased from zero","commit":"f4b0ad7","knative.dev/key":"default/example-00001","knative.dev/pod":"example-00001-deployment-678bd67c76-xd6w4"}
{"severity":"INFO","timestamp":"2021-09-04T00:19:09.445299796Z","logger":"queueproxy","caller":"queue/concurrency_state.go:52","message":"Requests dropped to zero","commit":"f4b0ad7","knative.dev/key":"default/example-00001","knative.dev/pod":"example-00001-deployment-678bd67c76-xd6w4"}

```

logs for pod-freezer:
```
{"severity":"INFO","timestamp":"2021-09-04T00:19:07.44293111Z","caller":"pod-freezer/main.go:34","message":"resume request received, thawing pod..."}
{"severity":"INFO","timestamp":"2021-09-04T00:19:07.444625075Z","caller":"pod-freezer/main.go:34","message":"resume request received, thawing pod..."}
{"severity":"INFO","timestamp":"2021-09-04T00:19:09.445622954Z","caller":"pod-freezer/main.go:32","message":"pause request received, freezing pod..."}
{"severity":"INFO","timestamp":"2021-09-04T00:19:09.447745441Z","caller":"pod-freezer/main.go:32","message":"pause request received, freezing pod..."}
```

So the pod-freezer service got two resume/pause requests, but QP only sent one of each....

This doesn't happen every time, and fwiw even when it does the end result seems to be what we want (i.e. the pod was resumed and then paused only after all requests finished). I won't be able to dig into it more until next week, but figured it was worth noting before then. 

It also could very well just be some issue with the pod-freezer and not QP, in which case we can pretend this comment was never made :smile: ...