@psschwei 2021-09-07
@julz just checked it, and yes the duplicate logs also occur in the current code (i.e. #11802)