@psschwei 2021-09-09
Still looking into the double logs, but interestingly, in testing today I saw several combinations:

* QP: 2 resume & 2 pause requests (RESUME-PAUSE-RESUME-PAUSE); freezer: 2 resume & 2 pause requests (RESUME-PAUSE-RESUME-PAUSE)
* QP: 1 resume and 1 pause (RESUME-PAUSE); freezer: 1 resume and 1 pause (RESUME-PAUSE)
* QP: 1 resume and 1 pause (RESUME-PAUSE); freezer: 2 resume and 2 pause (RESUME-RESUME-PAUSE-PAUSE)

The last case there is the issue I identified previously. So I decided to check the value of `inFlight` throughout the process. ([link to code](https://github.com/psschwei/serving/blob/rewrite-concurrencystate/pkg/queue/concurrency_state.go#L35-L48)). 

Here's some cobbled-together logs (sources are queue-proxy and freezer-daemon, and I've removed from fields from each record for readability):
```
{ "timestamp":"2021-09-09T21:13:52.713101711Z","caller":"queue/concurrency_state.go:63","message":"in returned function, inFlight = 1"}
{ "timestamp":"2021-09-09T21:13:52.717416011Z", "caller":"queue/concurrency_state.go:83","message":"Requests increased from zero",
{ "timestamp":"2021-09-09T21:13:52.717161957Z", "caller":"queue/concurrency_state.go:63","message":"in returned function, inFlight = 2"}
{ "timestamp":"2021-09-09T21:13:52.717289764Z", "caller":"queue/concurrency_state.go:63","message":"in returned function, inFlight = 3"}
{ "timestamp":"2021-09-09T21:13:52.719407499Z", "caller":"queue/concurrency_state.go:63","message":"in returned function, inFlight = 4"}
{ "timestamp":"2021-09-09T21:13:52.719672658Z","caller":"pod-freezer/main.go:34","message":"resume request received, thawing pod..."}
{ "timestamp":"2021-09-09T21:13:52.719991858Z","caller":"pod-freezer/main.go:34","message":"resume request received, thawing pod..."}
{ "timestamp":"2021-09-09T21:13:52.724756589Z", "caller":"queue/concurrency_state.go:67","message":"not paused"}
{ "timestamp":"2021-09-09T21:13:52.724876425Z", "caller":"queue/concurrency_state.go:67","message":"not paused"}
{ "timestamp":"2021-09-09T21:13:52.724696194Z", "caller":"queue/concurrency_state.go:67","message":"not paused"}
{ "timestamp":"2021-09-09T21:13:54.72707362Z", "caller":"queue/concurrency_state.go:46","message":"in deferred function, inFlight = 4"}
{ "timestamp":"2021-09-09T21:13:54.727604852Z", "caller":"queue/concurrency_state.go:46","message":"in deferred function, inFlight = 3"}
{ "timestamp":"2021-09-09T21:13:54.727971434Z", "caller":"queue/concurrency_state.go:46","message":"in deferred function, inFlight = 2"}
{ "timestamp":"2021-09-09T21:13:54.728228677Z", "caller":"queue/concurrency_state.go:46","message":"in deferred function, inFlight = 1"}
{ "timestamp":"2021-09-09T21:13:54.728259064Z", "caller":"queue/concurrency_state.go:53","message":"Requests dropped to zero"}
{ "timestamp":"2021-09-09T21:13:54.730455224Z","caller":"pod-freezer/main.go:32","message":"pause request received, freezing pod..."}
{ "timestamp":"2021-09-09T21:13:54.730971747Z","caller":"pod-freezer/main.go:32","message":"pause request received, freezing pod..."}

```